+++ .travis.yml
@@ -3,27 +3,24 @@ language: python
 cache: pip
 
 before_install:
-    sudo apt-get install -y python3-dev
+    - sudo apt-get install -y python3-dev
 
 install:
   - pip install --upgrade pip setuptools
-  - pip install virtualenv tox
+  - pip install virtualenv tox wheel
+  - tox --version
 
 script: tox -v
 
 matrix:
   include:
-  - python: "2.7"
-    env: TOXENV=py27
-
-  - python: "3.4"
-    env: TOXENV=py34
 
   - python: "3.6"
     env: TOXENV=checks
 
-  - python: "3.5"
-    env: TOXENV=py35
+  - python: "3.8"
+    dist: xenial
+    env: TOXENV=mypy
 
   - python: "3.6"
     env: TOXENV=py36
@@ -32,5 +29,22 @@ matrix:
     dist: xenial
     env: TOXENV=py37
 
-  - python: "pypy"
-    env: TOXENV=pypy
+  - python: "3.8"
+    dist: xenial
+    env: TOXENV=py38
+
+  - python: "3.9-dev"
+    dist: bionic
+    env: TOXENV=py39
+
+  - python: "nightly"
+    dist: bionic
+    env: TOXENV=py310
+
+  - python: "3.8"
+    dist: xenial
+    env: TOXENV=mypy
+
+jobs:
+  allow_failures:
+    - python: "nightly"
