+++ changelog.d/pr290.trivial.rst
@@ -1,0 +1,1 @@
+Add supported Python versions to :command:`black`.
