--- docs/_static/css/default.css
@@ -1,6 +1,0 @@
-/* Customize logo width */
-
-.wy-side-nav-search > a img.logo {
-	width: 6em;
-	background: white;
-}
