+++ docs/usage.rst
@@ -455,7 +455,7 @@ To compare two versions depends on your type:
     >>> v > dict(major=1, unknown=42)
     Traceback (most recent call last):
     ...
-    TypeError: __init__() got an unexpected keyword argument 'unknown'
+    TypeError: ... got an unexpected keyword argument 'unknown'
 
 
 Other types cannot be compared.
