+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.black]
 line-length = 88
-target-version = ['py37']
+target-version = ['py36', 'py37', 'py38']
 include = '\.pyi?$'
 # diff = true
 exclude = '''
