+++ setup.cfg
@@ -1,6 +1,6 @@
 [tool:pytest]
-norecursedirs = .git build .env/ env/ .pyenv/ .tmp/ .eggs/
-testpaths = . docs
+norecursedirs = .git build .env/ env/ .pyenv/ .tmp/ .eggs/ venv/
+testpaths = tests docs
 filterwarnings =
     ignore:Function 'semver.*:DeprecationWarning
 addopts =
@@ -16,6 +16,7 @@ max-line-length = 88
 ignore = F821,W503
 exclude =
     .env,
+    venv,
     .eggs,
     .tox,
     .git,
@@ -24,3 +25,15 @@ exclude =
     dist
     docs
     conftest.py
+
+[pycodestyle]
+count = False
+# ignore = E226,E302,E41
+max-line-length = 88
+statistics = True
+exclude =
+    .env,
+    .eggs,
+    .tox,
+    .git,
+    docs
