+++ tox.ini
@@ -1,11 +1,13 @@
 [tox]
 envlist =
     flake8
-    py{27,34,35,36,37}
-    pypy
+    py{36,37,38,39,310}
+    docs
+    mypy
+
 
 [testenv]
-description = Run test suite
+description = Run test suite for {basepython}
 whitelist_externals = make
 commands = pytest {posargs:}
 deps =
@@ -14,11 +16,13 @@ deps =
 setenv =
     PIP_DISABLE_PIP_VERSION_CHECK = 1
 
+
 [testenv:black]
 description = Check for formatting changes
 basepython = python3
 deps = black
 commands = black --check {posargs:.}
+
 
 [testenv:flake8]
 description = Check code style
@@ -26,11 +30,20 @@ basepython = python3
 deps = flake8
 commands = flake8 {posargs:}
 
+
+[testenv:mypy]
+description = Check code style
+basepython = python3
+deps = mypy
+commands = mypy {posargs:--ignore-missing-imports .}
+
+
 [testenv:docstrings]
 description = Check for PEP257 compatible docstrings
 basepython = python3
 deps = docformatter
 commands = docformatter --check {posargs:--pre-summary-newline semver.py}
+
 
 [testenv:checks]
 description = Run code style checks
@@ -38,10 +51,12 @@ basepython = python3
 deps =
     {[testenv:black]deps}
     {[testenv:flake8]deps}
+    {[testenv:mypy]deps}
     {[testenv:docstrings]deps}
 commands =
     {[testenv:black]commands}
     {[testenv:flake8]commands}
+    {[testenv:mypy]commands}
     {[testenv:docstrings]commands}
 
 [testenv:docs]
@@ -66,5 +81,5 @@ deps =
     wheel
     twine
 commands =
-    python3 setup.py sdist bdist_wheel --universal
+    python3 setup.py sdist bdist_wheel
     twine check dist/*
